+++ SocialDistribution/post/admin.py
@@ -2,6 +2,5 @@ from django.contrib import admin
 from . import models
 
 
-admin.site.register(models.POST)
-admin.site.register(models.Like)
+admin.site.register(models.Post)
 admin.site.register(models.Inbox)
